refactor(coral): Move StatsDisplay to main components folder
