+++ coral/src/app/components/StatsDisplay.tsx
